requirements.txt
numpy
waitress
flask
typing_extensions
torch
nemo_toolkit["asr"]
